Namespace cleanup.
