+++ Source/NexusForever.Game.Static/Reward/RewardRotationItemType.cs
@@ -1,10 +1,10 @@
-﻿namespace NexusForever.Game.Static.Rewards
+﻿namespace NexusForever.Game.Static.Reward
 {
     public enum RewardRotationItemType
     {
         None            = 0x0,
         Currency        = 0x1,
         Item            = 0x2,
-        AccountCurrency = 0x3,
+        AccountCurrency = 0x3
     };
 }
