+++ Source/NexusForever.Game.Static/Reward/RewardRotationRewardType.cs
@@ -1,10 +1,10 @@
-﻿namespace NexusForever.Game.Static.Rewards
+﻿namespace NexusForever.Game.Static.Reward
 {
     public enum RewardRotationRewardType
     {
         None     = 0x0,
         Item     = 0x1,
         Essence  = 0x2,
-        Modifier = 0x3,
+        Modifier = 0x3
     };
 }
