+++ Source/NexusForever.Game/Account/Entitlement/AccountEntitlementManager.cs
@@ -5,7 +5,7 @@ using NexusForever.Game.Abstract.Account;
 using NexusForever.Game.Abstract.Account.Entitlement;
 using NexusForever.Game.Entitlement;
 using NexusForever.Game.Static.Entity;
-using NexusForever.Game.Static.Rewards;
+using NexusForever.Game.Static.Reward;
 using NexusForever.GameTable;
 using NexusForever.GameTable.Model;
