+++ Source/NexusForever.Game/Account/Reward/RewardProperty.cs
@@ -1,7 +1,7 @@
 ﻿using NexusForever.Game.Abstract.Account.Reward;
-using NexusForever.Game.Static.Rewards;
+using NexusForever.Game.Static.Reward;
 using NexusForever.GameTable.Model;
-using NexusForever.Network.World.Message.Model.Rewards;
+using NexusForever.Network.World.Message.Model.Reward;
 
 namespace NexusForever.Game.Account.Reward
 {
