+++ Source/NexusForever.Network.World/Message/Model/Reward/ClientRewardRotationsRequestUpdate.cs
@@ -1,7 +1,7 @@
-﻿using NexusForever.Game.Static.Rewards;
+﻿using NexusForever.Game.Static.Reward;
 using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model
+namespace NexusForever.Network.World.Message.Model.Reward
 {
     [Message(GameMessageOpcode.ClientRewardRotationsRequestUpdate)]
     public class ClientRewardRotationsRequestUpdate : IReadable
