+++ Source/NexusForever.Network.World/Message/Model/Reward/ClientRewardTrackClaimRewardPoint.cs
@@ -1,6 +1,6 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model
+namespace NexusForever.Network.World.Message.Model.Reward
 {
     [Message(GameMessageOpcode.ClientRewardTrackClaimRewardPoint)]
     public class ClientRewardTrackClaimRewardPoint : IReadable
