+++ Source/NexusForever.Network.World/Message/Model/Reward/LockedReward.cs
@@ -1,7 +1,7 @@
-﻿using NexusForever.Game.Static.Rewards;
+﻿using NexusForever.Game.Static.Reward;
 using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.Rewards
+namespace NexusForever.Network.World.Message.Model.Reward
 {
     public class LockedReward : IWritable
     {
