+++ Source/NexusForever.Network.World/Message/Model/Reward/RewardTrack.cs
@@ -1,14 +1,14 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.Rewards
+namespace NexusForever.Network.World.Message.Model.Reward
 {
-    [Message(GameMessageOpcode.ServerRewardTrackUpdate)]
-    public class ServerRewardTrackUpdate : IWritable
+    public class RewardTrack : IWritable
     {
         public uint RewardTrackId { get; set; }
         public uint RewardPointsEarned { get; set; }
         public NetworkBitArray ClaimedRewards { get; set; } = new NetworkBitArray(32, NetworkBitArray.BitOrder.LeastSignificantBit);
         public uint Unlocked { get; set; } // needs more research, unclear if this is a bit array or an integer
+        public List<uint> RewardTrackRewardsIds { get; set; } = [];
         public bool IsActive { get; set; }
 
         public void Write(GamePacketWriter writer)
@@ -16,7 +16,9 @@ namespace NexusForever.Network.World.Message.Model.Rewards
             writer.Write(RewardTrackId, 14u);
             writer.Write(RewardPointsEarned);
             writer.WriteBytes(ClaimedRewards.GetBuffer());
-            writer.Write(Unlocked);
+            writer.Write(RewardTrackRewardsIds.Count);
+            writer.Write(Unlocked, 32u);
+            RewardTrackRewardsIds.ForEach(rewardId => writer.Write(rewardId));
             writer.Write(IsActive);
         }
     }
