+++ Source/NexusForever.Network.World/Message/Model/Reward/ServerRewardRotationLockedAdd.cs
@@ -1,9 +1,9 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.Rewards
+namespace NexusForever.Network.World.Message.Model.Reward
 {
-    [Message(GameMessageOpcode.ServerRewardRotationLockedRemove)]
-    public class ServerRewardRotationLockedRemove : IWritable
+    [Message(GameMessageOpcode.ServerRewardRotationLockedAdd)]
+    public class ServerRewardRotationLockedAdd : IWritable
     {
         public LockedReward Reward { get; set; }
