+++ Source/NexusForever.Network.World/Message/Model/Reward/ServerRewardRotationLockedList.cs
@@ -1,6 +1,6 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.Rewards
+namespace NexusForever.Network.World.Message.Model.Reward
 {
     [Message(GameMessageOpcode.ServerRewardRotationLockedList)]
     public class ServerRewardRotationLockedList : IWritable
