+++ Source/NexusForever.Network.World/Message/Model/Reward/ServerRewardRotations.cs
@@ -1,7 +1,7 @@
-﻿using NexusForever.Game.Static.Rewards;
+﻿using NexusForever.Game.Static.Reward;
 using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.Rewards
+namespace NexusForever.Network.World.Message.Model.Reward
 {
     [Message(GameMessageOpcode.ServerRewardRotations)]
     public class ServerRewardRotations : IWritable
