+++ Source/NexusForever.Network.World/Message/Model/Reward/ServerRewardTrackAdd.cs
@@ -1,6 +1,6 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.Rewards
+namespace NexusForever.Network.World.Message.Model.Reward
 {
     [Message(GameMessageOpcode.ServerRewardTrackAdd)]
     public class ServerRewardTrackAdd : IWritable
